NSU PlagiChecker2/backend/requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
pdfplumber
python-pptx
sentence-transformers
faiss-cpu
numpy
nltk
reportlab